address PR review round 1
